requirements.txt
Django==3.2.25
python-dotenv==1.0.1
PyMySQL==1.1.1
psycopg2-binary==2.9.9
